\name{datasets-JASPAR}
\alias{jaspar.insects.PFM}
\alias{jaspar.insects}
\docType{data}

\title{
  JASPAR database of Position Frequency Matrices and PWMs
}
\description{
  The JASPAR (http://jaspar.genereg.net/) database retrieved on April 2012. Contains the Position Frequency Matrices (PFMs) for
  a number of transcription factors. 
}

\usage{
data(jaspar.insects.PFM)
data(jaspar.insects)
}

\details{
	The \code{jaspar.insects.PFM} object contain a list of motifs in matrix format where rows are the four nucleotides (A,C,G,T). JASPAR identifiers
	have been stripped for clarity. 
	
	The PWMs (\code{jaspar.insects}) were generated by using the background frequencies of A,C,G,T in a set of 10031 2kb promoters in D. Melanogaster. 
}

\examples{
data(jaspar.insects.PFM)
data(jaspar.insects)
}
\keyword{datasets}
\seealso{
	 \code{datasets-bg-JASPAR}.
}
